+++ CHANGELOG.md
@@ -32,6 +32,7 @@ However, it is impossible to avoid those, especially in an earlier phase of deve
 
 ## [Unreleased]
 
+* :new: It is now possible to ask BaHaMAS to recreate the executable file(s) in the `continue` and `continue-thermalization` modes via the `--updateExecutable` option.
 * :new: It is now possible to use one time scale only for the MD integration specifying only one integer in the `i` field in the betas file.
 * :boom: Some setup variable names have been improved. This require running the `setup` mode once again to update the user-setup.
 * :white_check_mark: Improve functional tests to cover more real-life cases.